nvim(editor plugins): set optional flag
Because these specs are meant to override default options of plugins
that har managed by LazyVim. If I decide to temporarily disable some
LazyVim extras that install these plugins, I don't want to force them to
install if optional flag is not set. 

Setting optional to true has the effect that the plugin is not installed
if LazyVim (or I) do not specify this plugin elsewhere.